requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
pillow>=9.5.0
matplotlib>=3.7.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0 